Netflix Home Page
The Netflix Front Page Redesign project is a creative and ambitious GitHub project aimed at reimagining and enhancing the user interface (UI) and user experience (UX) of Netflix's streaming platform. Netflix is a popular streaming service, and this project seeks to provide a fresh, user-friendly, and visually appealing design for its front page.